+++ devicehive/gateway/tests/__init__.py
@@ -1,12 +1,12 @@
-# -*- encoding: utf8 -*-
-# vim: set et tabstop=4 shiftwidth=4 nu nowrap: fileencoding=utf-8 encoding=utf-8
+# -*- coding: utf-8 -*-
+# vim:set et tabstop=4 shiftwidth=4 nu nowrap fileencoding=utf-8 encoding=utf-8:
 
 
 import unittest
-from binary_tests import *
 
-__all__ = ['binary_tests']
 
 if __name__ == '__main__':
-    unittest.main()
-
+    runner = unittest.TextTestRunner()
+    tests = unittest.defaultTestLoader.discover('.')
+    for testsuite in tests :
+        runner.run(testsuite)
