+++ devicehive/gateway/xbeeapi.py
@@ -1,5 +1,5 @@
-# -*- encoding: utf8 -*-
-# vim: set et tabstop=4 shiftwidth=4 nu nowrap: fileencoding=utf-8 encoding=utf-8
+# -*- coding: utf-8 -*-
+# vim:set et tabstop=4 shiftwidth=4 nu nowrap fileencoding=utf-8 encoding=utf-8:
 
 import struct
